AI Agent/requirements.txt
groq>=0.4.1
python-dotenv>=1.0.0
streamlit>=1.37.0
